terra/pyproject.toml
[tool.poetry]
name = "terra"
version = "0.1.0"
description = "Django app that provisions DYMAX internal API and background tasks"
authors = ["Damián Silvani <munshkr@gmail.com>"]
license = "BSD-2-Clause"

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
Django = "3.0.5"
django-anymail = "7.2.1"
django-cors-headers = "3.5.0"
django-cron = "0.5.1"
django-guardian = "2.3.0"
django-jsoneditor = "0.1.6"
django-rest-auth = "0.9.5"
django-rq = "2.3.2"
django-storages = "1.10"
djangorestframework = "3.11.0"
djangorestframework-api-key = "2.0.0"
djangorestframework-gis = "0.15"
drf-yasg = "1.17.1"
filetype = "1.0.7"
google-api-python-client = "1.11.0"
google-auth-httplib2 = "0.0.4"
google-auth-oauthlib = "0.4.1"
google-cloud-bigquery = "1.27.2"
google-cloud-pubsub = "1.7.0"
google-cloud-storage = "1.31.0"
google-resumable-media = "1.0.0"
gunicorn = "20.0.4"
mailchimp3 = "3.0.14"
psycopg2-binary = "2.8.6"
pyproj = "^3.3.1"
python-dateutil = "2.8.1"
python-dotenv = "0.14.0"
requests = "2.24.0"
rq = "^1.10.1"
scikit-image = "^0.19.2"
sentry-sdk = "0.12.3"
Shapely = "1.7.1"
stripe = "2.51.0"
rasterio = "^1.2.10"
Fiona = "^1.8.21"
django-allauth = "0.42.0"
pytest = "^7.1.2"
django-mock-queries = "^2.1.7"
pytest-mock = "^3.8.2"
pytest-django = "^4.5.2"
pytest-dotenv = "^0.5.2"
drf-nested-routers = "^0.93.4"
setuptools = "^63.2.0"

[tool.poetry.dev-dependencies]
honcho = "^1.1.0"
black = "^22.1.0"
pre-commit = "^2.17.0"
ipython = "^8.1.1"
coverage = "^6.3.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"